requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.6
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pprintpp==0.4.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==6.3.0
pytest-mock==3.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
tinydb==4.8.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
